More clean-up
So much repeated code! Cooper is slimming down.
